Loofa Holder v1
Loofa holder first iteration: A cap to go on top of our home shower handle

Background: Our home shower is very minimilast and as a result has very little places to hang things. So our loofas can only hang on the handle that also turns on the water. When the water is turned on, the loofas always slide off onto the ground. This print is designed to prevent them from falling off.