Light Switch Protector v1
Light switch protector first iteration: A shield to cover part of a light switch to prevent it from getting wet.

Background: My girlfriend's apartment has the bathroom lightswitch dangerously close to where the shower water can spray. So for peace of mind for both of us, I made this shield. This makes it so the light switch doesn't get wet to prevent potential hazards.